requirements.txt
web3==7.6.0
py-solc-x==2.0.3